Move lib/posts/stats to state/posts/stats
