+++ client/post-editor/editor-slug/index.jsx
@@ -16,7 +16,7 @@ import classNames from 'classnames';
  */
 import TrackInputChanges from 'components/track-input-changes';
 import FormTextInput from 'components/forms/form-text-input';
-import { recordStat, recordEvent } from 'lib/posts/stats';
+import { recordStat, recordEvent } from 'state/posts/stats';
 import { getSelectedSiteId } from 'state/ui/selectors';
 import { getEditorPostId } from 'state/ui/editor/selectors';
 import { getEditedPostSlug } from 'state/posts/selectors';
